refac: move Version tests to own file
Version was previously refactored out of `promote.py`, this brings the
corresponding tests out to their own file too
